requirements.txt
streamlit
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langgraph==0.0.39
langchain-classic==0.0.1
python-dotenv
openai
pymupdf
pdf2image
pytesseract
Pillow
sentence-transformers
scikit-learn
numpy
sentencepiece
protobuf
# FastAPI Backend
fastapi
uvicorn[standard]
python-multipart
# Search Tools
tavily-python
duckduckgo-search